feat(training): kanal train, and a window derived rather than chosen
The job that was missing between "the gate exists" and "something is serving":
build a split, fit the candidates, run the gate, act on the verdict. One
function rather than a notebook, because every step has to happen identically
when it runs unattended.

Two judgement calls inside it, both pinned by tests:

The incumbent is whatever the registry says, not this run's best candidate. A
run that promoted its own winner would compare challengers against each other
and never against what is actually serving.

A run that promotes nothing exits zero. Treating "no promotion" as an error
teaches whoever watches CI to ignore red, and the gate exists precisely so
that most challengers do not pass.

The first real run was refused, and the refusal was right

  majority          macro-F1 0.0370   ECE 0.1024
  tfidf-linearsvc   macro-F1 0.3015   ECE 0.1007
  verdict: HOLD
    G7 calibration: ECE 0.1007 exceeds 0.08 (underconfident by 0.101)

Underconfidence is the wrong direction for a linear model, so the suspicion
was that fitting a temperature on 55 validation rows was making things worse.
Measured, and the suspicion was wrong:

  fitted on validation   T=0.558   ECE 0.1007
  no calibration         T=1.000   ECE 0.1609
  fitted on train        T=0.050   ECE 0.5262

Calibration helps. Even the best temperature in a sweep only reaches 0.0836,
just short of the gate. The model was genuinely not good enough on 163
training rows across 8 classes.

The window was the problem, and the fix is derived rather than picked

The age distribution says where the fault is:

  50th percentile   1.26 days
  75th percentile   3.35 days
  90th percentile  85.40 days     <- the evergreen cliff
  95th percentile 214.91 days

Three quarters of the corpus is under 3.4 days old and then it jumps to
months. The protocol's 14/7-day window measures the wrong side of that cliff,
because it was fixed when the assumption was weeks of collection history.

The fallback is derived from the operational question - classify the next
day's news given everything before it - which gives a one-day test window, a
one-day validation window, and training on the rest. No part of those numbers
refers to a score.

That distinction is load-bearing, because the windows were also swept and a
sweep is exactly what a protocol exists to stop anyone selecting from:

  train/val  train  test  macro-F1     ECE
     14/7      163  1077    0.3015  0.1007
      3/1      347   517    0.3451  0.0673
      2/1      470   517    0.4200  0.0683
      1/0.5    778   341    0.7102  0.0835

Those macro-F1 values are NOT comparable with each other. Narrowing the
window does not only hand the model more data - it makes the task easier,
because predicting twelve hours ahead is a different problem from predicting
seven days ahead. The horizon is part of the task definition, so every run
records horizon_days and prints a warning saying scores at one horizon cannot
be read against another. Anyone reading 0.71 next to 0.30 as an improvement
is being misled, and the run report now says so itself.

The default stays 14/7. The fallback fires only when the default produces
more test rows than training rows, and it logs loudly when it does.

There is now a champion

  split 0c6d397f63aa  train=470 val=308 test=517  horizon=1d
  tfidf-linearsvc  macro-F1 0.4200  ECE 0.0683  $0.000013/1k
  verdict: PROMOTE

Its per-class scores are recorded with it, which is what G4 compares against
on the next run, and they show the expected shape:

  internasional         0.710
  hiburan               0.594
  gaya-hidup-kesehatan  0.463
  ekonomi               0.144
  hukum-kriminal        0.083     <- nearly collapsed

hukum-kriminal is the smallest class and the one the protocol named when
choosing macro-F1 as the primary metric. It is behaving exactly as predicted,
and accuracy alone would have hidden it entirely.

Also: temporal_split now takes float day counts, since a corpus with days of
history needs sub-day windows; and the test fixture was resized after every
promotion test failed on n_test >= 150 - the gate working correctly against a
fixture that could never satisfy it.

347 tests.

+++ src/kanal/data/splits.py
@@ -152,8 +152,11 @@ def temporal_split(
     articles: list[Article],
     *,
     anchor: datetime | None = None,
-    train_cutoff_days: int = TRAIN_CUTOFF_DAYS,
-    val_cutoff_days: int = VAL_CUTOFF_DAYS,
+    # Float, not int. A corpus with only days of collection history needs
+    # sub-day windows, and the training job derives them from the prediction
+    # horizon rather than from the protocol's default.
+    train_cutoff_days: float = TRAIN_CUTOFF_DAYS,
+    val_cutoff_days: float = VAL_CUTOFF_DAYS,
 ) -> SplitManifest:
     """Split by publication time, assigning whole clusters.
